refactor(editor): Move useMessage into @n8n/design-system behind re-export shim (no-changelog)
Relocates the `useMessage` composable (CAT-3686 Stage 3 / N8N-37) from
editor-ui's `@/app/composables` into `@n8n/design-system`, the cycle-free home
approved for it: design-system already owns element-plus and N8nCheckbox, which
`useMessage` renders. The old `@/app/composables/useMessage` path becomes a thin
re-export shim (deprecation comment, no call-site codemod) so the ~90 existing
importers keep working during the migration.

- `sanitizeIfString` is now consumed from `@n8n/frontend-utils` — its actual
  post-merge home after N8N-36 (#34355), not `@n8n/composables` as the original
  plan assumed. design-system declares `@n8n/frontend-utils` as a dependency;
  frontend-utils depends only on vue + xss, so the edge is cycle-free.
- N8nCheckbox is now imported relatively (internal to the package).
- tsconfig `paths` and the vitest alias resolve `@n8n/frontend-utils` from source,
  mirroring the existing `@n8n/composables`/`@n8n/utils` setup.
- Adds a colocated unit test covering the composable's contract, HTML
  sanitization, and the confirmation-checkbox path.
