+++ packages/frontend/@n8n/design-system/package.json
@@ -63,6 +63,7 @@
   "dependencies": {
     "@internationalized/date": "catalog:frontend",
     "@n8n/composables": "workspace:*",
+    "@n8n/frontend-utils": "workspace:*",
     "@n8n/utils": "workspace:*",
     "@tanstack/vue-table": "^8.21.2",
     "@tiptap/core": "catalog:",
