+++ packages/frontend/@n8n/design-system/src/index.ts
@@ -9,6 +9,8 @@ export * from './directives';
 export type { IconOrEmoji } from './components/N8nIconPicker/types';
 export { IconBodyLoaderKey, useInjectIconBodyLoader } from './composables/useIconBodyLoader';
 export type { IconBodyLoader } from './composables/useIconBodyLoader';
+export { useMessage } from './composables/useMessage';
+export type { MessageBoxConfirmResult } from './composables/useMessage';
 export { default as N8nSelect2 } from './v2/components/Select/Select.vue';
 export { default as N8nSelect2Item } from './v2/components/Select/SelectItem.vue';
 export type * from './v2/components/Select/Select.types';
