+++ packages/frontend/@n8n/design-system/tsconfig.json
@@ -8,6 +8,7 @@
 		"paths": {
 			"@n8n/design-system*": ["./src*"],
 			"@n8n/composables*": ["../composables/src*"],
+			"@n8n/frontend-utils*": ["../frontend-utils/src*"],
 			"@n8n/utils*": ["../../../@n8n/utils/src*"]
 		}
 	},
