+++ packages/frontend/@n8n/design-system/vite.config.mts
@@ -40,6 +40,7 @@ export default mergeConfig(
 				'@': resolve(__dirname, 'src'),
 				'@n8n/design-system': resolve(__dirname, 'src'),
 				'@n8n/composables(.*)': resolve(packagesDir, 'frontend', '@n8n', 'composables', 'src$1'),
+				'@n8n/frontend-utils(.*)': resolve(packagesDir, 'frontend', '@n8n', 'frontend-utils', 'src$1'),
 				'@n8n/utils(.*)': resolve(packagesDir, '@n8n', 'utils', 'src$1'),
 			},
 		},
